feat: resolvers abstraction
